need to fix play again starts same instance of existing game

+++ lib/database.rb
@@ -1,13 +1,33 @@
 module Database
-  def save_game(game_name)
-    Dir.mkdir('saved') unless Dir.exist?('saved')
-    #yaml = YAML.dump(self)
-    filename = "../saved/#{game_name}.yaml"
-    File.open(filename, 'w') do |file|
-      file.write save_to_yaml
+  def save_game
+    puts 'Enter name for saved game'
+    # game_name = gets.chomp
+    # Dir.mkdir('saved') unless Dir.exist?('saved')
+    save_name = gets.chomp
+    file_path = File.expand_path("../saved/#{save_name}.yaml", __dir__)
+    return unless File.exist?(file_path)
+
+    puts "File #{save_name} already exists. Are you sure you want to overwrite the file? (Yes/No)"
+    overwrite_choice = gets.chomp.downcase
+    if overwrite_choice === 'n' || overwrite_choice ==='no'
+      handle_guess
+    elsif overwrite_choice === 'y' || overwrite_choice ==='yes'
+      filename = "../saved/#{save_name}.yaml"
+      File.open(filename, 'w') do |file|
+        file.write save_to_yaml
+      end
+      puts 'Your game has been saved. Thanks for playing!'
+      exit
+    else puts 'Please either type yes or no'
+      save_game
     end
-    puts 'Your game has been saved. Thanks for playing!'
-    exit
+
+    #       filename = "../saved/#{save_name}.yaml"
+    #     File.open(filename, 'w') do |file|
+    #       file.write save_to_yaml
+    #     end
+    #     puts 'Your game has been saved. Thanks for playing!'
+    #     exit
   end
 
   def save_to_yaml
@@ -20,19 +40,16 @@ module Database
     )
   end
 
-  #unserialise?
-
+  # unserialise?
 
   def file_list
     files = []
-    directory = File.expand_path("../../saved",__FILE__)
+    directory = File.expand_path('../saved', __dir__)
     puts "dir path: #{directory}"
     all_files = Dir["#{directory}/*"]
     puts "files in dir: #{all_files} end"
     Dir.entries(directory).each do |name|
-      if name.match(/\.yaml\z/)
-        files << name.gsub(/\.yaml\z/, "")
-      end
+      files << name.gsub(/\.yaml\z/, '') if name.match(/\.yaml\z/)
     end
     files
   end
@@ -43,35 +60,36 @@ module Database
     end
   end
 
-  def display_saved_games(number,name)
+=begin
+  def display_saved_games(number, name)
     <<~HEREDOC
-    \e[34m[#{number}]\e[0m #{name}
+      \e[34m[#{number}]\e[0m #{name}
     HEREDOC
   end
 
   def display_start_choice
-    text1 = "Play a new game"
-    text2 = "Load a saved game"
+    text1 = 'Play a new game'
+    text2 = 'Load a saved game'
     <<~HEREDOC
-    \e[34m[#{1}]\e[0m #{text1}
-    \e[34m[#{2}]\e[0m #{text2}
+      \e[34m[1]\e[0m #{text1}
+      \e[34m[2]\e[0m #{text2}
     HEREDOC
   end
+=end
 
   def load_saved_file
     puts "Here are the current saved games. Please choose which you'd like to load."
     show_file_list
-    puts "enter file name"
+    puts 'enter file name'
     @saved_game = gets.chomp
     file = YAML.safe_load(File.read("../saved/#{@saved_game}.yaml"))
+    puts "loading #{saved_game}..."
     @secret_word = file['secret word']
-     @placeholder = file['placeholder']
-      @victory =file['victory']
-      @guess_count = file['guess count']
-      @guess_array = file['guess array']
-      p @guess_array
-      p @placeholder
-      #puts "player enter your guess"
-      #@guess = gets.chomp
+    @placeholder = file['placeholder']
+    @victory = file['victory']
+    @guess_count = file['guess count']
+    @guess_array = file['guess array']
+    p @guess_array
+    p @placeholder
   end
 end
